patch from From: Jim Hayward <jimhayward@linuxexperience.com>
This patch adds keyboard mnemonics to all of the dialogs that did not
have them before. Quite a few of the dialogs still need some layout
redesigns I think. For now just the keyboard mnemonics and some minor
changes to a couple of the dialogs.

also some changes by Olivier to get the php icon back in place