On Mon, 2004-12-06 at 09:14 +1300, Alastair Porter wrote: > Some more things: > The find dialog only has the last 10 entries when you re-open a project. > The Anchor dialog has all entries - this should be cut back as well limit-url-list.patch Not sure if some other things should be limited as well, such as the classlist (also in the anchor dialog). would probably be a good idea > > Opening a project with no files in it opens the project properly, but > doesn't set the window title to show "project name" - "filename" - > "bluefish" project-title.patch
>
> Alastair
>
> On Sun, 2004-12-05 at 16:59 +1300, Alastair Porter wrote:
> > Just looking through some things that could/should be looked at before
> > 1.0
> >
> > Enter doesn't do 'ok' in the save dialog
file-dialog-default-action.patch

> > Extra directories added to the left pane in the open/dave dialogs - i
> > really don't like these, it makes it feel cluttered
file-dialog-remove-extra-directories.patch
just commented out, not removed.
I got the impression that this functionality was just for specalist
directories, such as a desktop publishing program adding a clipart
location.

> >
> > The funcref tree has alternating stripes on it (this is an option on the
> > treeview). The bookmakrs, filebrowser and funcref should all match each
> > other
left-pane-{add|remove}-rules.patch
after adding rules to all the other trees, i thought that some of them
didn't really look great with them, such as the file list in the 2 pane
filebrowser. removing them from the funcref doesn't look too bad.
The comment in the function reference
(http://developer.gimp.org/api/2.0/gtk/GtkTreeView.html#gtk-tree-view-set-rules-hint)
says:
Do not use it just because you prefer the appearance of the ruled tree;
that's a question for the theme. Some themes will draw tree rows in
alternating colors even when rules are turned off, and users who prefer
that appearance all the time can choose those themes. You should call
this function only as a semantic hint to the theme engine that your tree
makes alternating colors useful from a functional standpoint (since it
has lots of columns, generally).

So was it used in the funcref window to make it easier to use, or just
from a visual point of view