new specfile Have done some changes in the file list in order to the match latest changes in CVS.
The desktop icon patch is not required anymore and was removed.
The specfile should go to CVS.

The bluefish icon for *.bfproject is right shown now, but a click on
it opens the project file instead of the project.

>Shall I better change this to --enable-devel-install or
>> --enable-unstable-install?

I think '--enable-unstable-install'. I don't use this configure option
currently anymore with my specfile.

The changed specfile is attached, should go to CVS.

Currently no probs anymore with click on *.bfproject on Fedora Core 3.