Fixed issue reported by Hylke van der Schaaf: During make install, the reference xml files are copied to /usr/local/share/bluefish/ and THEN gzipped there.
This means that for every file there, gzip gives a warning that the gz
file allready exists.
It also means that both the gzipped and the non gzipped version are
installed.

Wouldn't it be much nicer to make the gzip process run during make (as
non root) and then only install the gzipped files with make install?