Re-implement working GUI language choosing. Not all languages are given in native language, because I don't know/find them.
Idea: Maybe we can use gettext() for this approach. Use it like this: define a list of languages. Now loop over it and do setlocale(lang), gettext(nativelang) -> add this to the list of languages. So we don't need the struct and translators will care about the correct native language name.